# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://laravel-api.test",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of posts",
                "description": "Returns list of posts",
                "operationId": "getPostsList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Store a new post",
                "description": "Create a new post",
                "operationId": "storePost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a specific post by ID",
                "description": "Returns a single post",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of post to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update an existing post",
                "description": "Updates a post and returns the updated post",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of post to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status",
                                    "category_id",
                                    "feature_image_url",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Post Title"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "feature_image_url": {
                                        "type": "string",
                                        "example": "http://example.com/image.jpg"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description of the post."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete an existing post",
                "description": "Deletes a post and returns a success message",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of post to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Post Delete Successful"
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePostRequest": {
                "title": "Store Post Request",
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "title": {
                        "description": "The title of the post",
                        "type": "string",
                        "example": "Example Post Title"
                    },
                    "feature_image_url": {
                        "description": "URL of the feature image",
                        "type": "string",
                        "example": "http://example.com/feature-image.jpg"
                    },
                    "status": {
                        "description": "Status of the post",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "The category ID associated with the post",
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "description": "The description of the post",
                        "type": "string",
                        "example": "This is an example description of the post."
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "title": "Post",
                "required": [
                    "title",
                    "slug",
                    "status",
                    "category_id",
                    "category_type"
                ],
                "properties": {
                    "id": {
                        "description": "The post ID",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "The title of the post",
                        "type": "string",
                        "example": "Example Post Title"
                    },
                    "slug": {
                        "description": "The slug of the post",
                        "type": "string",
                        "example": "example-post-title"
                    },
                    "feature_image_url": {
                        "description": "URL of the feature image",
                        "type": "string",
                        "example": "http://example.com/feature-image.jpg"
                    },
                    "status": {
                        "description": "Status of the post",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "The category ID associated with the post",
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "description": "The description of the post",
                        "type": "string",
                        "example": "This is an example description of the post."
                    },
                    "created_at": {
                        "description": "The time the post was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "The time the post was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}